move albumTile into section folder
