+++ app/javascript/app/Content/Section/Carousel/AlbumTile/LoadingTile/LoadingTile.svelte
